+++ gmft/_rich_text/__init__.py
@@ -1,0 +1,3 @@
+"""
+Note: this module is experimental and may be subject to refactors and changes.
+"""