+++ gmft/common.py
@@ -1,3 +1,6 @@
+from typing import TypeVar, Union
+
+
 class Rect:
     """
     A floating-point rectangle.
@@ -58,3 +61,4 @@ class Rect:
         return (self.bbox[2] - self.bbox[0]) * (self.bbox[3] - self.bbox[1])
     
 Rect.EMPTY = Rect((0, 0, 0, 0))
+