+++ gmft/detectors/img2table.py
@@ -1,6 +1,6 @@
 
 import copy
-from dataclasses import dataclass, fields
+from dataclasses import dataclass
 from typing import Generator
 import img2table.document
 import numpy as np
@@ -9,8 +9,7 @@ import pandas as pd
 from gmft.common import Rect
 from gmft.detectors.common import BaseDetector, CroppedTable
 from gmft.formatters.common import FormattedTable
-from gmft.pdf_bindings.common import BasePDFDocument, BasePage, _infer_line_breaks
-from gmft.utils import auto_init
+from gmft.pdf_bindings.common import BasePDFDocument, BasePage
 
 
 try:
@@ -27,7 +26,7 @@ from img2table.tables.objects.extraction import ExtractedTable
 
 class Img2TablePDFDocument(BasePDFDocument, img2table.document.PDF):
     """
-    Wraps a basePDFdocument in the img2table format
+    Wraps a BasePDFdocument in the img2table format
     """
     
     _underlying: BasePDFDocument = None
@@ -111,7 +110,7 @@ class Img2TablePDFDocument(BasePDFDocument, img2table.document.PDF):
 
 class Img2TablePage(img2table.document.PDF):
     """
-    Wraps a BasePage as a singleton document in the img2table format.
+    Wraps a BasePage as a singleton document in the img2table format, 
     because detectors work on a page level.
     """
     
@@ -350,7 +349,8 @@ class Img2TableTable(FormattedTable):
         """
         raise NotImplementedError("This is not implemented, instead create Tables using Img2TableDetector.extract")
 
-@auto_init
+# @auto_init
+@dataclass
 class Img2TableDetectorConfig:
     implicit_rows: bool = False
     implicit_columns: bool = False
@@ -378,10 +378,15 @@ class Img2TableDetector(BaseDetector[Img2TableDetectorConfig]):
         else:
             i2tpage = Img2TablePage(page)
         
-        config = self.config
         if config_overrides is not None:
-            config = copy.deepcopy(self.config)
-            config.__dict__.update(config_overrides.__dict__)
+            if isinstance(config_overrides, dict):
+                # only update if it's a dict
+                pass # TODO
+            else:
+                # override everything
+                config = config_overrides
+        else:
+            config = self.config
         
         extracted_tables = i2tpage.extract_tables(ocr=None,
                                                   **config.__dict__)