+++ gmft/formatters/tatr.py
@@ -1,6 +1,9 @@
 
 import copy
+from dataclasses import dataclass, field
+from typing import Union
 
+from gmft._dataclasses import non_defaults_only, with_config
 from gmft.detectors.common import CroppedTable, RotatedCroppedTable
 from gmft.formatters.common import FormattedTable, TableFormatter, _normalize_bbox
 from gmft.pdf_bindings.common import BasePage
@@ -11,8 +14,8 @@ from gmft.table_function_algorithm import extract_to_df
 from gmft.table_visualization import plot_results_unwr
 
 
-
-class TATRFormatterConfig:
+@dataclass
+class TATRFormatConfig:
     """
     Configuration for :class:`.TATRTableFormatter`.
     """
@@ -43,7 +46,7 @@ class TATRFormatterConfig:
     (having fewer rows than expected merges cells, which is bad).
     """
     
-    cell_required_confidence = {
+    cell_required_confidence: dict = field(default_factory=lambda: {
         0: 0.3, # table
         1: 0.3, # column
         2: 0.3, # row
@@ -51,7 +54,7 @@ class TATRFormatterConfig:
         4: 0.5, # projected row header
         5: 0.5, # spanning cell
         6: 99   # no object
-    }
+    })
     """Confidences required (>=) for a row/column feature to be considered good. See TATRFormattedTable.id2label
     
     But low confidences may be better than too high confidence (see formatter_base_threshold)
@@ -61,19 +64,19 @@ class TATRFormatterConfig:
     
     # ---- options ----
     
-    remove_null_rows = True 
+    remove_null_rows: bool = True 
     """remove rows with no text"""
     
-    enable_multi_header = False
+    enable_multi_header: bool = False
     """Enable multi-indices in the dataframe.
     If false, then multiple headers will be merged column-wise."""
     
-    semantic_spanning_cells = False
+    semantic_spanning_cells: bool = False
     """
     [Experimental] Enable semantic spanning cells, which often encode hierarchical multi-level indices.
     """
     
-    semantic_hierarchical_left_fill = 'algorithm'
+    semantic_hierarchical_left_fill: Union[str, None] = 'algorithm'
     """
     [Experimental] When semantic spanning cells is enabled, when a left header is detected which might
     represent a group of rows, that same value is reduplicated for each row.
@@ -86,56 +89,56 @@ class TATRFormatterConfig:
     
     # ---- large table ----
     
-    large_table_if_n_rows_removed = 8
+    large_table_if_n_rows_removed: int = 8
     """
     If >= n rows are removed due to non-maxima suppression (NMS), then this table is classified as a large table.
     """
     
-    large_table_threshold = 10
+    large_table_threshold: int = 10
     """with large tables, table transformer struggles with placing too many overlapping rows
     luckily, with more rows, we have more info on the usual size of text, which we can use to make
     a guess on the height such that no rows are merged or overlapping
     
     large table assumption is only applied when (# of rows > large_table_threshold) AND (total overlap > large_table_row_overlap_threshold)
     set 9999 to disable, set 0 to force large table assumption to run every time"""
-    large_table_row_overlap_threshold = 0.2
+    large_table_row_overlap_threshold: float = 0.2
     
-    large_table_maximum_rows = 1000
+    large_table_maximum_rows: int = 1000
     """If the table predicts a large number of rows, refuse to proceed. Therefore prevent memory issues for super small text."""
     
-    force_large_table_assumption=None
+    force_large_table_assumption: Union[bool, None]=None
     """True: force large table assumption to be applied to all tables
     False: force large table assumption to not be applied to any tables
     None: heuristically apply large table assumption according to threshold and overlap"""
 
     # ---- rejection and warnings ----
 
-    total_overlap_reject_threshold = 0.2
+    total_overlap_reject_threshold: float = 0.2
     """reject if total overlap is > 20% of table area"""
     
-    total_overlap_warn_threshold = 0.05
+    total_overlap_warn_threshold: float = 0.05
     """warn if total overlap is > 5% of table area"""
     
-    nms_warn_threshold = 5
+    nms_warn_threshold: int = 5
     """warn if non maxima suppression removes > 5 rows"""
     
-    iob_reject_threshold = 0.05
+    iob_reject_threshold: float = 0.05
     """reject if iob between textbox and cell is < 5%"""
 
-    iob_warn_threshold = 0.5
+    iob_warn_threshold: float = 0.5
     """warn if iob between textbox and cell is < 50%"""
     
     # ---- technical ----
     
-    _nms_overlap_threshold = 0.1
+    _nms_overlap_threshold: float = 0.1
     """Non-maxima suppression: if two rows overlap by > threshold (default: 10%), then the one with the lower confidence is removed.
     A subsequent technique is able to fill in gaps created by NMS."""
     
-    _large_table_merge_distance = 0.6
+    _large_table_merge_distance: float = 0.6
     """In the large_table method, if two means are within (60% * text_height) of each other, then they are merged.
     This may be useful to adjust if text is being split due to subscripts/superscripts."""
     
-    _smallest_supported_text_height = 0.1
+    _smallest_supported_text_height: float = 0.1
     """The smallest supported text height. Text smaller than this height will be ignored. 
     Helps prevent very small text from creating huge arrays under large table assumption."""
     
@@ -172,11 +175,7 @@ class TATRFormatterConfig:
     def deduplication_iob_threshold(self, value):
         raise DeprecationWarning("deduplication_iob_threshold is deprecated. See nms_overlap_threshold instead.")
     
-    def __init__(self, torch_device: str = None):
 
-        if torch_device is not None:
-            self.torch_device = torch_device
-    
 
 class TATRFormattedTable(FormattedTable):
     """
@@ -200,7 +199,7 @@ class TATRFormattedTable(FormattedTable):
     label2id = {v: k for k, v in id2label.items()}
     
     
-    config: TATRFormatterConfig
+    config: TATRFormatConfig
     outliers: dict[str, bool]
     
     effective_rows: list[tuple]
@@ -224,17 +223,17 @@ class TATRFormattedTable(FormattedTable):
     
     def __init__(self, cropped_table: CroppedTable, fctn_results: dict, 
                 #  fctn_scale_factor: float, fctn_padding: tuple[int, int, int, int], 
-                 config: TATRFormatterConfig=None):
+                 config: TATRFormatConfig=None):
         super(TATRFormattedTable, self).__init__(cropped_table)
         self.fctn_results = fctn_results
         # self.fctn_scale_factor = fctn_scale_factor
         # self.fctn_padding = tuple(fctn_padding)
         
         if config is None:
-            config = TATRFormatterConfig()
+            config = TATRFormatConfig()
         self.config = config
         self.outliers = None
-    def df(self, recalculate=False, config_overrides: TATRFormatterConfig=None):
+    def df(self, recalculate=False, config_overrides: TATRFormatConfig=None):
         """
         Return the table as a pandas dataframe. 
         :param recalculate: by default, the dataframe is cached
@@ -243,11 +242,8 @@ class TATRFormattedTable(FormattedTable):
         if recalculate == False and config_overrides is None and self._df is not None: # cache
             return self._df
         
-        if config_overrides is not None:
-            config = copy.deepcopy(self.config)
-            config.__dict__.update(config_overrides.__dict__)
-        else:
-            config = self.config
+        config = with_config(self.config, config_overrides)
+
         
         self._df = extract_to_df(self, config=config)
         return self._df
@@ -318,7 +314,7 @@ class TATRFormattedTable(FormattedTable):
         return {**parent, **{
             # 'fctn_scale_factor': self.fctn_scale_factor,
             # 'fctn_padding': list(self.fctn_padding),
-            'config': self.config.__dict__,
+            'config': non_defaults_only(self.config),
             'outliers': self.outliers,
             'fctn_results': self.fctn_results,
         }, **optional}
@@ -329,17 +325,15 @@ class TATRFormattedTable(FormattedTable):
         Deserialize from dict.
         A page is required partly because of memory management, since having this open a page may cause memory issues.
         """
+        d = copy.deepcopy(d) # don't modify the original dict
         cropped_table = CroppedTable.from_dict(d, page)
         
         if 'fctn_results' not in d:
             raise ValueError("fctn_results not found in dict -- dict may be a CroppedTable but not a TATRFormattedTable.")
         
-        config = TATRFormatterConfig()
-        for k, v in d['config'].items():
-            if v is not None and config.__dict__.get(k) != v:
-                setattr(config, k, v)
+        config = TATRFormatConfig(**d['config'])
         
-        results = d['fctn_results']
+        results = d['fctn_results'] # fix shallow copy issue
         if 'fctn_scale_factor' in d or 'scale_factor' in d or 'fctn_padding' in d or 'padding' in d:
             # deprecated: this is for backwards compatibility
             scale_factor = d.get('fctn_scale_factor', d.get('scale_factor', 1))
@@ -363,12 +357,12 @@ class TATRFormatter(TableFormatter):
     Using :meth:`extract`, a :class:`.FormattedTable` is produced, which can be exported to csv, df, etc.
     """
     
-    def __init__(self, config: TATRFormatterConfig=None):
+    def __init__(self, config: TATRFormatConfig=None):
         import transformers
         from transformers import AutoImageProcessor, TableTransformerForObjectDetection
         
         if config is None:
-            config = TATRFormatterConfig()
+            config = TATRFormatConfig()
         if not config.warn_uninitialized_weights:
             previous_verbosity = transformers.logging.get_verbosity()
             transformers.logging.set_verbosity(transformers.logging.ERROR)
@@ -388,11 +382,8 @@ class TATRFormatter(TableFormatter):
         Extract the data from the table.
         """
         
-        if config_overrides is not None:
-            config = copy.deepcopy(self.config)
-            config.__dict__.update(config_overrides.__dict__)
-        else:
-            config = self.config
+        config = with_config(self.config, config_overrides)
+
         
         image = table.image(dpi=dpi, padding=padding, margin=margin) # (20, 20, 20, 20)
         padding = table._img_padding
@@ -433,6 +424,5 @@ class TATRFormatter(TableFormatter):
 
 # legacy aliases from the nonstandard days
 TATRTableFormatter = TATRFormatter
-TATRFormatConfig = TATRFormatterConfig