+++ test/test_imports.py
@@ -1,0 +1,18 @@
+
+# test to_dict and from_dict
+
+# test all legacy imports
+
+
+import pytest
+import gmft
+from gmft.pdf_bindings import PyPDFium2Document
+
+    
+
+def test_imports():
+    # these locations:
+    # gmft.* 
+    # gmft.table_function
+    # gmft.table_detection
+    pass